Move GQL Requests to Java Enum
- Alphabetize GQL Enum
- Add new GQL Requests: DELETE_SCHEDULING_GOAL, GET_EFFECTIVE_MODEL_ARGUMENTS, DELETE_SIMULATION_PRESET
- Remove Deprecated Request: GET_EFFECTIVE_ARGUMENTS
- Changed INSERT_PROFILE_SEGMENT into INSERT_PROFILE_SEGMENTS
- Get additional fields in: GET_PLAN, GET_EXTERNAL_DATASET
- Remove extra parameter in GET_SCHEDULING_DSL_TYPESCRIPT
